feat: move http client to separated directory
