+++ mod.ts
@@ -4,5 +4,6 @@ export * from "./src/script/templates/index.js"
 export * from "./src/transaction/index.js"
 export * from "./src/transaction/fee-models/index.js"
 export * from "./src/transaction/broadcasters/index.js"
+export * from "./src/transaction/http/index.js"
 export * from "./src/messages/index.js"
 export * from "./src/compat/index.js"
